Fix implicit import resolution with multiple folder-based Python modules
Previously, the import resolver failed to handle multiple folder-based
Python modules (those with __init__.py), when resolving implicit
imports, due to storing only filenames in the `HashMap`. This caused
naming conflicts when there was multiple folder-based Python modules.

To fix we store both the directory and file name in the `HashMap`.