Refactor how symbol references are computed
Previously, references of a symbol were collected during the
`SymbolTable` construction. This had a limitation which we
weren't able to get references of a symbol if it was part of
an attribute expression.

Example, `foo.bar` if `foo` could bea module or a class that
was defined in another module, since the `SymbolTable`
construction step is only aware of the the current file is
building the symbol table for, we couldn't get the proper
information to compute the symbol reference.

Now, the symbol references are computed on-demand outside of
`SymbolTable` by the `ReferencesFinder`.