Add the stub and non-stub path to import declarations
Previously, we only stored one path in the import declaration,
it could be a path to a stub file or the non-stub file. Now,
we store both of them so we can do proper analysis on it.

The indexer, when indexing the collected imports, for now,
if both stub and non-stub file are present, gives preference
for indexing the stub file. The non-stub file won't be indexed.
This may change in the future.