Remove `resolve_declaration` from `SymbolTableDb`
This function was causing numerous bugs, and after careful
consideration, I concluded it was unnecessary. Its original purpose was
to traverse import chains to locate the source declaration of a symbol.
While this process occurred during the computation of completion item
candidates, the results were only intended for use in the
`completionItem/resolve` request method, which displays documentation
for the selected completion item. We will now perform this computation
directly within the `completionItem/resolve` handler.