Fix unresolved import diagnostic not showing for repeated imports with the same symbol
Previously, the symbol table only created a single declaration for
imports with the same symbol. Example:
```python
from foo.bar import Bar
from foo import Foo
```
In this example we would only have one declaration created for the first
`foo`. This approach had the side-effect that if we failed to resolve
the import of `foo`, we would only see in the LSP diagnostics the first
`foo` being highlighted.

Now, we create a special-case declaration for this situation
(`DeclStmt::SameImport`), that points to the first `foo` declaration
made. This way we have the necessary information to properly compute the
diagnostics.