Replace `String::from_utf8` with `String::from_utf8_lossy` in `get_python_version` and `get_python_search_paths` functions
Previoulsy, we were using `String::from_utf8` to parse the output of the
python command to a string and unwraping the value with `expect`, if the
string wasn't valid UTF-8 the function would panic and the server would
crash.

We replace it with the non-panicking function `String::from_utf8_lossy`.
This function replaces invalid UTF-8 characters with a special symbol,
preventing the server from crashing.