Explicitly visit f-string during symbol table construction
When building the `NodeStack` we explicitly visit f-string to add it's
nodes to the stack, we were not doing the same in the
`SymbolTableBuilder` causing the two to be out of sync. When pushing
nodes in one of these two we MUST do the same in the other. Otherwise,
this can cause very subtle and weird bugs.