Fix not showing references in other files if the symbol was part of an attr expression and it was trigged inside a function
We have a heuristic to avoid doing a global search for references across
the indexed files, for symbols that were defined locally, e.g, inside a
function. But this doesn't apply if the symbol is part of an attribute
expression, e.g, the symbol could be a member variable of a class
defined in another file.